+++ src/extension.ts
@@ -27,12 +27,14 @@ import {
   LanguageClient,
   LanguageClientOptions,
   ServerOptions,
+  URI,
 } from "vscode-languageclient/node";
 import { getHubSrc } from "./hub";
 import { RuleInputProvider } from "./rule-search";
 import { ScanResultProvider } from "./rule-search-results";
 import { ChatProvider, ChatRequest } from "./chat";
 import { Recipe, RecipeProvider } from "./recipes";
+import { askSourceryCommand } from "./ask-sourcery";
 
 function createLangServer(): LanguageClient {
   const token = workspace.getConfiguration("sourcery").get<string>("token");
@@ -215,62 +217,10 @@ function registerCommands(
     })
   );
 
-  // {'message': {'type': 'recipe_request', 'data': {'id': 'debug_code', 'name': 'Debug', 'kind': 'recipe_request'}}
-
-  context.subscriptions.push(
-    commands.registerCommand("sourcery.chat.ask", () => {
-      showAskSourceryQuickPick(recipeProvider.recipes).then((result: any) => {
-        vscode.commands.executeCommand("sourcery.chat.focus").then(() => {
-          let request;
-          if ("id" in result) {
-            request = {
-              type: "recipe_request",
-              data: {
-                kind: "recipe_request",
-                name: result.label,
-                id: result.id,
-              },
-            };
-          } else {
-            request = {
-              type: "chat_request",
-              data: { kind: "user_message", message: result.label },
-            };
-          }
-
-          vscode.commands.executeCommand("sourcery.chat_request", request);
-        });
-      });
-    })
-  );
-
   context.subscriptions.push(
     commands.registerCommand("sourcery.chat.ask", (arg?) => {
-      showAskSourceryQuickPick(recipeProvider.recipes).then((result: any) => {
-        vscode.commands.executeCommand("sourcery.chat.focus").then(() => {
-          let request;
-          let contextRange = "start" in arg ? arg : null;
-          if ("id" in result) {
-            request = {
-              type: "recipe_request",
-              data: {
-                kind: "recipe_request",
-                name: result.label,
-                id: result.id,
-              },
-              context_range: contextRange,
-            };
-          } else {
-            request = {
-              type: "chat_request",
-              data: { kind: "user_message", message: result.label },
-              context_range: contextRange,
-            };
-          }
-
-          vscode.commands.executeCommand("sourcery.chat_request", request);
-        });
-      });
+      let contextRange = arg && "start" in arg ? arg : null;
+      askSourceryCommand(recipeProvider.recipes, contextRange);
     })
   );
 
@@ -456,6 +406,8 @@ function registerCommands(
     commands.registerCommand(
       "sourcery.chat_request",
       (message: ChatRequest) => {
+        // Use the editor selection unless a range was passed through in
+        // the message
         let selectionLocation = getSelectionLocation();
         if (message.context_range != null) {
           selectionLocation = {
@@ -463,19 +415,7 @@ function registerCommands(
             range: message.context_range,
           };
         }
-        const activeEditor = window.activeTextEditor;
-        let activeFile = undefined;
-        if (activeEditor) {
-          activeFile = activeEditor.document.uri;
-        }
-        const allFiles = [];
-        for (const tabGroup of vscode.window.tabGroups.all) {
-          for (const tab of tabGroup.tabs) {
-            if (tab.input instanceof vscode.TabInputText) {
-              allFiles.push(tab.input.uri);
-            }
-          }
-        }
+        let { activeFile, allFiles } = activeFiles();
 
         let request: ExecuteCommandParams = {
           command: "sourcery/chat/request",
@@ -657,32 +597,19 @@ function openDocument(document_path: string) {
   });
 }
 
-function showAskSourceryQuickPick(recipes: Recipe[]) {
-  return new Promise((resolve) => {
-    const recipeNames = recipes.map((item) => item.name);
-    const recipeItems = recipes.map((item) => ({
-      label: item.name,
-      id: item.id,
-    }));
-
-    const quickPick = window.createQuickPick();
-    quickPick.placeholder = "Ask any question or choose one of these recipes";
-    quickPick.items = recipeItems;
-
-    quickPick.onDidAccept(() => {
-      const selection = quickPick.activeItems[0];
-      resolve(selection);
-      quickPick.hide();
-    });
-
-    quickPick.onDidChangeValue(() => {
-      // add what the user has typed to the pick list as the first item
-      if (!recipeNames.includes(quickPick.value)) {
-        const newItems = [{ label: quickPick.value }, ...recipeItems];
-        quickPick.items = newItems;
+function activeFiles(): { activeFile: URI | null; allFiles: URI[] } {
+  const activeEditor = window.activeTextEditor;
+  let activeFile = null;
+  if (activeEditor) {
+    activeFile = activeEditor.document.uri;
+  }
+  const allFiles = [];
+  for (const tabGroup of vscode.window.tabGroups.all) {
+    for (const tab of tabGroup.tabs) {
+      if (tab.input instanceof vscode.TabInputText) {
+        allFiles.push(tab.input.uri);
       }
-    });
-    quickPick.onDidHide(() => quickPick.dispose());
-    quickPick.show();
-  });
+    }
+  }
+  return { activeFile, allFiles };
 }
